EventHandler fix 1/2 way done. I won't tidy up the Config until tomorrow after I finish the Server part of the GUI. Now that I know why past attempts have failed involving the server part, (EventHandlers being the problem), I can do the server part now, and then include all of the GemsCraft Alpha 1.0.3.9 features.
Also included in Beta 1.0 will be better CPE implemetation.

Because CPE is handled differently on other fCraft platforms, I will not
include it in the Config.xml. I will have created a seperate file called
CPE.json for that, as to not cause the weird unknown config key issue with
other fCraft platforms.

Signed-off-by: Alex Potter <alex364981@gmail.com>