ConfigGUI redesign in progress > General Config is finished, ready to go > Rank Config is in progress, not finished yet, SEVERAL errors based on location of objects and controls. Still working through those. > Still working on migrating Rank Config Event Handlers. > Note: The Rank Config seems to be the most difficult one to switch because of how dynamic the content is.
I did, however, add a "ConfigHandler" class to help the Config Forms to connect with one another and to be able to access methods that more than one needs access to.

Also please note that the GUI design will not be completed and cleaned up until AFTER all the Configurations are setup completely. The icky designs and weird names and tabs are remaining until this part of the project is complete.

Signed-off-by: apotter96 <apotter96@godispower.com>